Fix SSA/ASS primary color {\1c} formatting
Formatting text with {\1c} would cause the end of the text line to
unconditionally contain a closing </font> tag. If the closing {\1c}
was included it would remain in the text and the text would be
suffixed with <font> regardless of the position of the closing {\1c}
e.g.: start {\1c}text{\1c} end => start <font>text{\1c} end</font>.
Add AdvancedSubStationAlpha and SubStationAlpha tests.