Adds ability to set layer for the current subtitle.
When manually positioning a lot of subtitles, which subtitle is rendered on top can often be incorrect, so a layer input would be useful.
The layer property in Paragraph is a 32-bit integer so the numeric up/down control accepts all 32-bit integers to be able to display any possible value.