Refactor code in FixCommonErrors form
Remove unnecessary condition checking in FixCommonErrors form. Certain redundant conditions were previously set in the form which checked whether a key already existed in _allowedFixes. However, List.Add does not throw an error if an existing key is added, and as such, this condition is unnecessary and was removed to simplify the code.