Refactor line break removal using new RemoveRecursiveLineBreaks method
This commit replaces the previous method of removing repeated line breaks in the text with a new RemoveRecursiveLineBreaks method. The previous implementation consisted of a loop that executed a Replace method each time a repeated line break was found in the text. This was not an optimal solution as it resulted in unnecessary overhead due to the continuous replacement process.

Now, the RemoveRecursiveLineBreaks method performs this operation more efficiently, without the need to continuously call the Replace method. The codebase has been updated to use the new method, improving execution speed and efficiency. Unit tests have also been added to confirm the functionality of this new method.