Add StringBuilderPool class in Common lib
The new StringBuilderPool class has been added to provide an implementation of an object pool pattern specifically for StringBuilders. This object pool assists with the use and re-use of StringBuilders, where more-often used string builders are returned from the pool and the least used ones are dropped. Limitations on pool size and capacity of string builders are also set to manage resource usage.