Update string extension tests for sentence endings
This commit updates the unit tests in StringExtensionsTest.cs to account for more cases of sentence endings. The original tests did not account for edge cases such as having an italic closing tag or an em dash before the end of the sentence which can also signify the end of a sentence. Additionally, corrected an erroneous test that claimed an empty string has a sentence ending, which it doesn't.