Refactor NikseComboBox to enhance null safety
The condition of null-checking before removing the existing list view and invalidating the form has been updated. This has been done in order to avoid potential null reference exceptions. This change adds an extra layer of null-safety by storing the result of `FindForm()` in a variable and then checking if the form is not null before making the subsequent operations on the form.