Add StringBuilderPoolTest to Test.csproj and implement tests
A new file, StringBuilderPoolTest.cs, has been added to the Test project to implement unit tests for the StringBuilderPool class. The tests cover various behaviours such as initial capacity assertion, returning to pool, and not returning to pool in case the capacity is larger than the maximum or the pool is already filled. This change will aid in maintaining the robustness of the StringBuilderPool class.