Update method names for font retrieval
Updated the method names in FontHelper.cs and its references in multiple files for more accurate description of the returned object. Changed 'GetBoldAndItalicCapableFonts' and 'GetItalicOrBoldCapableFonts' to 'GetBoldAndItalicCapableFontFamilies' and 'GetItalicOrBoldCapableFontFamilies' respectively, as they are returning FontFamily objects not Font objects. This provides clearer understanding of the method's functionality to other developers.