Update font helper queries to include regular fonts
Modified the FontHelper methods GetBoldAndItalicCapableFontFamilies and GetItalicOrBoldCapableFontFamilies in order to yield fonts that are capable of regular and bold styles, instead of just italic or bold. This change was made across multiple files where these methods are called. The adjustment allows for a greater range of fonts to be used, especially in instances where an italic style is not necessary or desired.