Refactor plugin loading logic and improve validation checks.
The refactoring was done for LocalPluginMetadataProvider and LocalPlugin classes. A new SymbolReader class is introduced to read property values from the plugin assemblies' types using reflection, removing repeated code for this task. String constants are introduced for the "*.dll" extension and for the properties names to avoid hardcoded strings.
In LocalPlugin, a validation check method 'IsValid' is introduced to ensure the plugin's name is not null or empty and its version is greater than 0. This is to ensure only valid plugins get added to the installed plugins. Improved the equality check for plugin names to ignore casing. All these changes made the code more maintainable and robust, reducing possibility for errors.