Refactor plugin update checker for better testability
Replaced concrete metadata providers with interfaces ILocalPluginMetadataProvider and IOnlinePluginMetadataProvider in PluginUpdateChecker. Constructor overloading is used to maintain the original behavior along with the new behavior where the PluginUpdateChecker accepts the mentioned interfaces as arguments. This change allows better unit testing.

Additional packages were added to the project for the creation of substitute objects. Unit tests were written to test new and old behavior of PluginUpdateChecker.

The "No newline at end of file" error was also corrected.