Add IsAllUppercase method and related tests
Added a new method IsAllUppercase in the StringExtensions class. This method checks whether all letter-characters in a given string are uppercase. Created test methods for this new functionality in the StringExtensionsTest class. Also, a new class FormattableTextReader with Read() method is added to read the text correctly in IsAllUppercase method. FormattableTextReaderText class is added for testing FormattableTextReader class. The main purpose of these changes is to provide a way to validate if the string is in uppercase, ignoring HTML tags or any non-letter characters.