Refactor OcrFixEngine to utilize FixExtraSpaces method
A section of code in OcrFixEngine has been refactored to improve readability and reduce repetition. Previously, two distinct while loops were used to remove additional spaces. This has been replaced by a call to the existing FixExtraSpaces method, which accomplishes the same goal in a simpler and more maintainable way.