Rename and update UnknownSubtitle25 class to MpSub
The class 'UnknownSubtitle25' has been renamed to 'MpSub' to reflect the identification of the previously unnamed subtitle format. This change is based on the specifications defined in the Mplayer's documentation. The name field of the class has also been updated to "MpSub" from "Unknown 25". The link to the relevant documentation has been commented in for reference.