Adds a few optimizations to pruning.
Caches coverage calculations to avoid iterating over the coverage records again.

The results depend on the subtitles currently visible, but it sometimes cuts prune time from around 70 ms to around 35 ms (about 50% faster) and sometimes has little or no effect (66 ms to 50 ms) or even a bit slower (from 70 ms to 72 ms). It is extremely rare to see this, though. It almost always has a noticeable speed gain from caching.

Also keep in mind that this is with the caching code executing first. If the runtime is caching the DisplaySubtitleHelper code, the non-caching version would already be at an advantage but it is still slower.

This is still in progress and needs cleaning up.