Minor cache improvement and boundary condition checks.
Instead of throwing away cache entries when adding new paragraphs, update them with the newly selected paragraph so they can be re-used.

In testing, this improved selection speed by about 5-7 percent (but with a standard deviation of up to 19 percent so there is quite a bit of variation here, sometimes making performance worse).