Add support for TimeLineAscii in 'ToText' method
Added support for exporting subtitles to TimeLineAscii format in SubtitleFormat class. This would help to provide different formats for the subtitles depending on the user's preference. A new internal class called 'TimeLineAsciiTextWriter' has been introduced to handle the writing of subtitle data into the TimeLineAscii format. There's also a change in the placement of TimeLineAscii in the formats list and a 'TODO' comment has been added for future reference.