Refactor logic for uppercase letter checking in French sentences
The code checking for uppercase letters following apostrophes in French sentences has been refactored and enhanced. The method `IsToKeepCasing` was renamed to `IsToKeepLUppercase` to improve clarity. An additional check, `IsWordAfterApostropheUppercase`, was added to ensure accurate casing for words following an apostrophe. A helpful method comment was also added for future reference.

Signed-off-by: Ivandro Jao <ivandrofly@gmail.com>