Refactor code for XML document creation
The XML document creation has been refactored into a separate function to reduce repetition. Prior code for creating the XML document was repeated in several places, specifically in the 'Remove' and 'Add' functions. The new 'CreateDocument' function now contains this code, making the overall codebase cleaner and easier to maintain.

Signed-off-by: Ivandro Jao <ivandrofly@gmail.com>