Refactor NameList to simplify addition of names
The logic for adding names to either _namesMultiList or _namesList in the NameList class was spread widely in the method. It has now been encapsulated within a new private method called TryAdd. This commit leads to cleaner and clearer code within the class and promotes easier maintenance in the future.

Signed-off-by: Ivandro Jao <ivandrofly@gmail.com>