Update async calls to use ConfigureAwait(false)
The code updates are primarily related to the usage of async calls across multiple translation services files. The async calls have been adjusted to use ConfigureAwait(false) for performance reasons and to avoid deadlocks. Changes are also made to several methods to make them asynchronously and thus improve the responsiveness of the operations.

Signed-off-by: Ivandro Jao <Ivandrofly@gmail.com>