Refactor duration calculations to use new Duration struct
Replaced usages of DurationTotalMilliseconds with a new Duration struct that better encapsulates the concept of a duration. This refactoring simplifies much of the code and also makes the code dealing with duration calculations more intuitive and easier to understand. The Duration struct includes methods for converting duration to various time units and to a TimeCode object.

Signed-off-by: Ivandro Jao <Ivandrofly@gmail.com>