Remove redundant dictionary update loops
Removed multiple unnecessary loops that were attempting to add entries to dictionaries if they do not already exist. This streamlines the dictionary update process and improves code readability. The logic for adding entries that do not exist has been deemed superfluous, as it was not contributing to the required functionality.

Signed-off-by: Ivandro Jao <Ivandrofly@gmail.com>